# Loosen keyword capture token criteria