--- a/src/compiler/build_tables/lex_table_builder.h
+++ b/src/compiler/build_tables/lex_table_builder.h
@@ -4,8 +4,21 @@
 #include <memory>
 #include <vector>
 #include <unordered_map>
-#include <set>
+#include <unordered_set>
+#include <utility>
+#include "compiler/parse_table.h"
 #include "compiler/lex_table.h"
+
+namespace std {
+
+using tree_sitter::rules::Symbol;
+
+template <>
+struct hash<pair<Symbol::Index, Symbol::Index>> {
+  size_t operator()(const pair<Symbol::Index, Symbol::Index> &) const;
+};
+
+} // namespace std
 
 namespace tree_sitter {
 
@@ -30,12 +43,23 @@ enum ConflictStatus {
   ),
 };
 
+struct CoincidentTokenIndex {
+  std::unordered_map<
+    std::pair<rules::Symbol::Index, rules::Symbol::Index>,
+    std::unordered_set<ParseStateId>
+  > entries;
+
+  bool contains(rules::Symbol, rules::Symbol) const;
+  const std::unordered_set<ParseStateId> &states_with(rules::Symbol, rules::Symbol) const;
+};
+
 class LexTableBuilder {
  public:
   static std::unique_ptr<LexTableBuilder> create(const SyntaxGrammar &,
                                                  const LexicalGrammar &,
                                                  const std::unordered_map<rules::Symbol, LookaheadSet> &,
-                                                 const std::vector<LookaheadSet> &);
+                                                 const CoincidentTokenIndex &,
+                                                 ParseTable *);
 
   struct BuildResult {
     LexTable main_table;
@@ -43,7 +67,7 @@ class LexTableBuilder {
     rules::Symbol keyword_capture_token;
   };
 
-  BuildResult build(ParseTable *);
+  BuildResult build();
 
   ConflictStatus get_conflict_status(rules::Symbol, rules::Symbol) const;