### Background

In #137, I added an optimization that tries to identify a set of tokens as 'keywords' and handles them via a separate tokenizer function. This significantly improved the compile time for most parsers, but it did not apply to all parsers.

### Solution

This PR generalizes the logic for identifying and extracting these 'keywords'. 

* The tokens no longer must be specified as single strings, which is helpful for cases like C's [`primitive_type`](https://github.com/tree-sitter/tree-sitter-c/blob/ac27b4ca2acd6c5f8f98c7228c73a6cc18b62939/grammar.js#L291-L306), in which a single token can match many several possible strings.
* I no longer enforce that the keyword capture token must match every possible keyword. I just find the token which matches the largest set of keywords. This allows keyword extraction to work for [`tree-sitter-rust`](https://github.com/tree-sitter/tree-sitter-rust), whereas before it bailed out.
* I've refined the logic for anticipating the lexical conflicts that could arise from substituting the keyword capture token. I now look at each individual state that's changing to determine if any conflicts would be introduced. This allows keyword extraction to work for [`tree-sitter-haskell`](https://github.com/tree-sitter/tree-sitter-haskell) where before it bailed out.

### Results



parser | old compile time | new compile time 
-------|------------------|-------------------
c | 2.6 s | 1.6 s
rust | 5.7 s | 2.2 s
haskell | 59.0 s | 22.6 s

/cc @rewinfrey - we briefly paired on trying to fix this. I finally got around to finishing it.